rename Go SDK key-value directory to key_value
